Created a default layout and a link data file
The link data file will contain an array of all the links to be added to
the main navigation bar of the website. The entries in the data file
will automatically be added to the navigation bar. The default layout contains
the head tag (and the encompassing body that contains the content variable)