Added open source licenses page
The open source licences contains a list of links to repositories of
software that I am using for my project (which are all stored in the
oslicenses data file). Since I plan to make this software free, I needed
to make an html table to let LibreJS know what software I plan on using.