Made the serve target to be PHONY; all has prereqs
This will make their prerequisite to be a special target (known as
PHONY). This target ensures that the recipe will still run whether or
not there is a file with the same name as the target in the current
directory.
The all target now has prerequisites properly specified so that it will
only run when necessary.