Simplified footer.md
The footer.md, although short, was hard to read. I moved all the links
and text into a yaml data file making easier to change.

However, I replaced the images that I was fetching from icon8 to
text. The text has the same properties as the links in the main (top)
navbar.